// apps/sales-copilot/src/generated/models/task-model.ts
import type { Account } from './account-model';
import type { Opportunity } from './opportunity-model';

export const TaskPrioritykeyToLabel = {
  'Prioritykey0': '高',
  'Prioritykey1': '中',
  'Prioritykey2': '低'
} as const;
export type TaskPrioritykey = keyof typeof TaskPrioritykeyToLabel;

export const TaskStatuskeyToLabel = {
  'Statuskey0': '待办',
  'Statuskey1': '进行中',
  'Statuskey2': '已完成',
  'Statuskey3': '已取消'
} as const;
export type TaskStatuskey = keyof typeof TaskStatuskeyToLabel;

export interface Task {
  /**
   * @displayName Task
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Title
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName Account
   */
  account?: Pick<Account, 'id' | 'name1'>;
  /**
   * @displayName Created On
   */
  createdon?: string;
  /**
   * @displayName Due Date
   * @validationRule Required for create/update operations
   */
  duedate: string;
  /**
   * @displayName Is Overdue
   */
  isoverdue?: boolean;
  /**
   * @displayName Notes
   */
  notes?: string;
  /**
   * @displayName Opportunity
   */
  opportunity?: Pick<Opportunity, 'id' | 'name1'>;
  /**
   * @displayName Overdue Reason
   */
  overduereason?: string;
  /**
   * @displayName Owner ID
   * @validationRule Required for create/update operations
   */
  ownerid: string;
  /**
   * @displayName Priority
   */
  priorityKey?: TaskPrioritykey;
  /**
   * @displayName Status
   */
  statusKey?: TaskStatuskey;
}

export const _Task = 'Task' as const;